@@ -164,7 +164,7 @@
     "@types/node-forge": "^1.3.1",
     "@types/node-schedule": "^2.1.0",
     "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
+    "@types/react-dom": "^18.0.10",
     "@types/react-router": "^5.1.17",
     "@types/react-router-dom": "^5.3.2",
     "@types/react-test-renderer": "^17.0.1",
@@ -180,7 +180,7 @@
     "css-loader": "^6.5.0",
     "css-minimizer-webpack-plugin": "^3.1.1",
     "detect-port": "^1.3.0",
-    "electron": "^18.0.4",
+    "electron": "^22.0.2",
     "electron-builder": "23.0.8",
     "electron-devtools-installer": "^3.2.0",
     "electron-notarize": "^1.1.1",
@@ -245,11 +245,11 @@
     "node-forge": "^1.3.1",
     "node-schedule": "^2.1.0",
     "oauth-1.0a": "^2.2.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-i18next": "^11.16.7",
     "react-icons": "^4.3.1",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.6.2",
     "recharts": "^2.1.16",
     "regenerator-runtime": "^0.13.9",
     "request": "^2.88.2"
