@@ -13,7 +13,7 @@ import {
 } from 'react-icons/fi';
 import log from 'electron-log';
 import { Link } from 'react-router-dom';
-import { Redirect } from 'react-router';
+import { Navigate } from 'react-router';
 import { useTranslation } from 'react-i18next';
 import { useEnvironmentList } from '../contexts/EnvironmentListContext';
 import { useNotifications } from '../contexts/NotificationsContext';
@@ -263,7 +263,7 @@ export default function CreateEnvironmentView(): JSX.Element {
         });
 
         updateEnvironmentList();
-        setValidationMessage(<Redirect to="/" />);
+        setValidationMessage(<Navigate to="/" />);
       } else if (
         permissionsResults.some(
           (i: { httpStatus: number }) =>
