@@ -1,7 +1,7 @@
 import log from 'electron-log';
 import { ipcRenderer } from 'electron';
 import { Suspense } from 'react';
-import { render } from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { I18nextProvider } from 'react-i18next';
 import { HashRouter } from 'react-router-dom';
 import App from './App';
@@ -27,13 +27,15 @@ log.transports.file.fileName = ipcRenderer.sendSync('getIsDevelopment')
   : 'fluig-monitor.log';
 log.transports.file.format = ipcRenderer.sendSync('getLogStringFormat');
 
-render(
-  <Suspense fallback="Loading">
-    <I18nextProvider i18n={i18n}>
-      <HashRouter>
-        <App />
-      </HashRouter>
-    </I18nextProvider>
-  </Suspense>,
-  document.getElementById('root')
-);
+const container = document.getElementById('root');
+if (container) {
+  createRoot(container).render(
+    <Suspense fallback="Loading">
+      <I18nextProvider i18n={i18n}>
+        <HashRouter>
+          <App />
+        </HashRouter>
+      </I18nextProvider>
+    </Suspense>
+  );
+}
