@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { useState } from 'react';
 import { motion } from 'framer-motion';
-import { Route, Switch, useParams, useRouteMatch } from 'react-router';
+import { Route, Routes, useParams } from 'react-router';
 import { useTranslation } from 'react-i18next';
 import {
   FiAirplay,
@@ -13,7 +13,6 @@ import {
 } from 'react-icons/fi';
 import { Link } from 'react-router-dom';
 import globalContainerVariants from '../utils/globalContainerVariants';
-import EnvironmentViewParams from '../../common/interfaces/EnvironmentViewParams';
 import EnvironmentSummary from '../components/EnvironmentDataView/EnvironmentSummary';
 import EditEnvironmentSettingsView from './EditEnvironmentSettingsView';
 
@@ -21,34 +20,32 @@ import '../assets/styles/views/EnvironmentView.scss';
 
 export default function EnvironmentView(): JSX.Element {
   const { t } = useTranslation();
-  const { environmentId }: EnvironmentViewParams = useParams();
-  // path is used to create paths relative to the parent, while url is used to create links
-  const { path, url } = useRouteMatch();
+  const { environmentId } = useParams();
   const [menuIsClosed, setMenuIsClosed] = useState(true);
 
   const submenuItems = [
     {
-      target: `${url}/summary`,
+      target: `summary`,
       text: t('views.EnvironmentDataContainer.sideMenu.summary'),
       icon: <FiAirplay />,
     },
     {
-      target: `${url}/database`,
+      target: `database`,
       text: t('views.EnvironmentDataContainer.sideMenu.database'),
       icon: <FiDatabase />,
     },
     {
-      target: `${url}/detailedMemory`,
+      target: `detailedMemory`,
       text: t('views.EnvironmentDataContainer.sideMenu.detailedMemory'),
       icon: <FiLayers />,
     },
     {
-      target: `${url}/artifacts`,
+      target: `artifacts`,
       text: t('views.EnvironmentDataContainer.sideMenu.artifacts'),
       icon: <FiPackage />,
     },
     {
-      target: `${url}/users`,
+      target: `users`,
       text: t('views.EnvironmentDataContainer.sideMenu.users'),
       icon: <FiUsers />,
     },
@@ -93,9 +90,9 @@ export default function EnvironmentView(): JSX.Element {
             </div>
             <div className="last-menu-items">
               <Link
-                to={`${url}/edit`}
-                onClick={() => setSelectedButton(`${url}/edit`)}
-                className={selectedButton === `${url}/edit` ? 'active' : ''}
+                to="edit"
+                onClick={() => setSelectedButton(`edit`)}
+                className={selectedButton === `edit` ? 'active' : ''}
               >
                 <FiSettings />{' '}
                 <span className="item-text">
@@ -107,43 +104,60 @@ export default function EnvironmentView(): JSX.Element {
         </section>
 
         <section id="menu-content">
-          <Switch>
-            <Route path={`${path}/summary`}>
-              <EnvironmentSummary environmentId={Number(environmentId)} />
-            </Route>
-            <Route path={`${path}/database`}>
-              <>
-                <h2>{t('views.EnvironmentDataContainer.sideMenu.database')}</h2>
-                <p>{t('components.global.underDevelopment')}</p>
-              </>
-            </Route>
-            <Route path={`${path}/detailedMemory`}>
-              <>
-                <h2>
-                  {t('views.EnvironmentDataContainer.sideMenu.detailedMemory')}
-                </h2>
-                <p>{t('components.global.underDevelopment')}</p>
-              </>
-            </Route>
-            <Route path={`${path}/artifacts`}>
-              <>
-                <h2>
-                  {t('views.EnvironmentDataContainer.sideMenu.artifacts')}
-                </h2>
-                <p>{t('components.global.underDevelopment')}</p>
-              </>
-            </Route>
-            <Route path={`${path}/users`}>
-              <>
-                <h2>{t('views.EnvironmentDataContainer.sideMenu.users')}</h2>
-                <p>{t('components.global.underDevelopment')}</p>
-              </>
-            </Route>
+          <Routes>
+            <Route
+              path="summary"
+              element={
+                <EnvironmentSummary environmentId={Number(environmentId)} />
+              }
+            />
+            <Route
+              path="database"
+              element={
+                <>
+                  <h2>
+                    {t('views.EnvironmentDataContainer.sideMenu.database')}
+                  </h2>
+                  <p>{t('components.global.underDevelopment')}</p>
+                </>
+              }
+            />
+            <Route
+              path="detailedMemory"
+              element={
+                <>
+                  <h2>
+                    {t(
+                      'views.EnvironmentDataContainer.sideMenu.detailedMemory'
+                    )}
+                    <p>{t('components.global.underDevelopment')}</p>
+                  </h2>
+                </>
+              }
+            />
+            <Route
+              path="artifacts"
+              element={
+                <>
+                  <h2>
+                    {t('views.EnvironmentDataContainer.sideMenu.artifacts')}
+                    <p>{t('components.global.underDevelopment')}</p>
+                  </h2>
+                </>
+              }
+            />
+            <Route
+              path="users"
+              element={
+                <>
+                  <h2>{t('views.EnvironmentDataContainer.sideMenu.users')}</h2>
+                  <p>{t('components.global.underDevelopment')}</p>
+                </>
+              }
+            />
 
-            <Route path={`${path}/edit`}>
-              <EditEnvironmentSettingsView />
-            </Route>
-          </Switch>
+            <Route path="edit" element={<EditEnvironmentSettingsView />} />
+          </Routes>
         </section>
       </div>
     </motion.div>
